move delete card function out of card.js
