+++ src/components/MyCards.js
@@ -1,12 +1,37 @@
+import axios from "axios";
+import { confirmAlert } from 'react-confirm-alert';
+
 import Card from "./Card";
 
 export default function MyCards(props) {
 
   const deleteCard = (id) => {
-    let newCards = props.myCards.filter(card => {
-      return card.id !== id;
+    confirmAlert({
+      title: 'Delete Card',
+      message: 'Are you sure to do this.',
+      buttons: [
+        {
+          label: 'Confirm',
+          onClick: () => {
+            const headers = {
+              headers: { Authorization: `Bearer ${props.currentUser.token}`}
+            };
+            return axios.delete(`http://localhost:8001/api/cards/${id}`, headers)
+            .then(r => {
+              console.log(r.data)
+              let newCards = props.myCards.filter(card => {
+                return card.id !== id;
+              });
+              props.setMyCards(newCards)
+            })
+            .catch(e => console.log(e))
+          }
+        },
+        {
+          label: 'Cancel',
+        }
+      ]
     });
-    props.setMyCards(newCards)
   }
 
   const editCard = (id) => {
@@ -20,7 +45,6 @@ export default function MyCards(props) {
       <Card
         key={card.id}
         card={card}
-        currentUser={props.currentUser}
         deleteCard={deleteCard}
         editCard={editCard}
       />
